% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between.R
\name{between}
\alias{between}
\title{Betweenness Centrality Calculation}
\usage{
between(A.mat, weight = T)
}
\arguments{
\item{A.mat}{An n x n adjacency matrix.}

\item{weight}{A boolean indicating if edges are weighted}
}
\value{
Returns a length-\code{n} vector of \code{c.betw.std}
}
\description{
\code{between} calculates the geodesics between two nodes
}
\examples{
A.mat = sim_adjacency(10)
between(A.mat)

}
