% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeness.R
\name{closeness}
\alias{closeness}
\title{Closeness Centrality Computation}
\usage{
closeness(A.mat)
}
\arguments{
\item{A.mat}{An n x n adjacency matrix.}
}
\value{
Returns a length-n vector of \code{c.closeness.std}
of standardized closeness centrality scores
}
\description{
\code{closeness} Closeness centrality measures how many steps is
required to access every other node from a given node.
}
\examples{
A.mat.w = sim_adjacency(10, weight = c(1, 10))
closeness(A.mat.w)

}
