% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{degree}
\alias{degree}
\title{Degree Centrality Computation}
\usage{
degree(A.mat, type = "undirected")
}
\arguments{
\item{A.mat}{An n x n adjacency matrix.}

\item{type}{A character string indicating the type of degree
centrality: "undirected", "indegree", or "outdegree".
By default, it is "undirected".}
}
\value{
Returns the value of \code{c.degree.std}, a length-\code{n} vector
of degree centrality scores
}
\description{
\code{degree} calculates the standardized degree centrality of the network;
a higher degree centrality indicates the node has more connections to
other nodes
}
\examples{
A.mat = sim_adjacency(10, mode = "directed")  # generate digraph
degree(A.mat)                            # consider both in and out degrees
degree(A.mat, type = "indegree")         # only count indegree

}
