% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's Algorithm}
\usage{
dijkstra(A.mat, src)
}
\arguments{
\item{A.mat}{An n x n adjacency matrix.}

\item{src}{A length 1 vector indicating the start node}
}
\value{
Returns a list of \code{dist} and \code{paths}
}
\description{
\code{dijkstra} calculates the geodesics between two nodes
}
\examples{
A.mat = sim_adjacency(10)
dijkstra(A.mat, src = 1)

}
