% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{geodesic.jik}
\alias{geodesic.jik}
\title{Helper function: count all geodesics passing through jk}
\usage{
geodesic.jik(paths, j, k, i = NULL)
}
\arguments{
\item{paths}{A length-\code{n} list of paths.}

\item{j}{Positive integer; node 1}

\item{k}{Positive integer; node 2}

\item{i}{Positive integer; in-between node. Default is \code{NULL}}
}
\value{
A list of two non-negative integers: \code{jk} is the number of geodesics
between node j and node k; \code{jik} is the number of number of these geodesics
passing through node i
}
\description{
\code{geodesic.jik} counts geodesics between two nodes
}
\examples{
A.mat = sim_adjacency(10)
geodesic.jik(A.mat)

}
