% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_centrality.R
\name{graph_centrality}
\alias{graph_centrality}
\title{Graph the network with centrality score}
\usage{
graph_centrality(
  graph,
  centr_score,
  main = "Centrality Network",
  seed = 1,
  legend = T
)
}
\arguments{
\item{graph}{A graph object as defined in the \code{igraph} package}

\item{centr_score}{A length-\code{n} vector of pre-calculated centrality score for each node}

\item{main}{A custom string. Title of the plot.}

\item{seed}{A seed value that can be custom specified. Default is 1.}

\item{legend}{A boolean specifying if we want to display the legend or not}
}
\value{
Function has no return value. It makes a cute plot though.
}
\description{
\code{graph_centrality} graphs the network (as a graph object) overlapped with the
specified centrality score. This function only performs the graphing and not
any computation
}
\examples{
A.mat = sim_adjacency(10, mode = "undirected")  # symmetric; all edge weights 1
A.graph = igraph::graph_from_adjacency_matrix(A.mat, mode = "undirected")
c.deg.und = degree(A.mat, type = "undirected")  # calculate centrality score

graph_centrality(A.graph, c.deg.und, main = "Degree Centrality (undirected)")

}
