% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_adjacency.R
\name{sim_adjacency}
\alias{sim_adjacency}
\title{Adjacency matrix simulation}
\usage{
sim_adjacency(n, mode = "undirected", weight = NULL)
}
\arguments{
\item{n}{Positive integer, number of nodes in the network}

\item{mode}{A character string, indicating directedness of the matrix.
Default is "undirected", where edges do not have direction (therefore adjacency
matrix symmetric)}

\item{weight}{A length-2 vector of non-negative integers, specifying the range
edge weights. Default is \code{NULL}, indicating unweighted edges.}
}
\value{
Returns an n x n adjacency matrix \code{A.mat}
}
\description{
\code{sim_adjacency} creates a randomly generated adjacency matrix
}
\examples{
sim_adjacency(10, mode = "undirected")   # generates a 10x10 symmetric matrix
with edges 1 and non-edges 0

sim_adjacency(10, mode = "directed", weight = c(10, 20))  # generates
a 10x10 matrix (not necessarily symmetric) with edge weights from a uniform
(10, 20) distribution.

}
