% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_adjacency.R
\name{sim_deg_conn}
\alias{sim_deg_conn}
\title{Adjacency matrix simulation -- varying connectedness}
\usage{
sim_deg_conn(n, p = 0.5)
}
\arguments{
\item{n}{Positive integer, number of nodes in the network}

\item{p}{Positive double, probability of a node having a connection with another.
Default is \code{0.5}.}
}
\value{
Returns an n x n adjacency matrix \code{A.mat}
}
\description{
\code{sim_deg_conn} creates a randomly generated adjacency matrix
with different degree of connectedness
}
\examples{
sim_deg_conn(10) # edge exists between 2 nodes with probability 0.5
sim_deg_conn(10, p = 0.1) # edge exists between 2 nodes with probability 0.1

}
