% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeness.R
\name{closeness}
\alias{closeness}
\title{Closeness Centrality Computation}
\usage{
closeness(A.mat, weight = T)
}
\arguments{
\item{A.mat}{An n x n adjacency matrix.}

\item{weight}{A boolean indicating if edges are weighted}
}
\value{
Returns the value of \code{c.closeness.std}
}
\description{
\code{closeness} calculates the closeness centrality of the network
}
\examples{

myfunction(1) # returns 1

}
