% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{degree}
\alias{degree}
\title{Degree Centrality Computation}
\usage{
degree(A.mat, type = "undirected")
}
\arguments{
\item{A.mat}{An n x n adjacency matrix.}

\item{type}{A character string indicating the type of degree
centrality: "undirected", "indegree", or "outdegree".
By default, it is "undirected".}
}
\value{
Returns the value of \code{c.degree.std}
}
\description{
\code{degree} calculates the degree centrality of the network
}
\examples{

myfunction(1) # returns 1

}
