% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenvect.R
\name{eigenvect}
\alias{eigenvect}
\title{Eigenvector Centrality Calculation}
\usage{
eigenvect(A.mat, weight = T)
}
\arguments{
\item{A.mat}{An n x n adjacency matrix.}

\item{weight}{A boolean indicating if edges are weighted}
}
\value{
Returns a vector \code{c.eig.std}
}
\description{
\code{eigenvect} calculates the Eigenvector Centrality of the Graph
}
\examples{
A.mat = sim_adjacency(10)
eigenvect(A.mat)

}
